requirements.txt
tube_dl
eyed3
ffmpeg-python